Forms: create shared ConsentToggle component
